Make index page a tsx component
